+++ src/Symfony/Component/DependencyInjection/Tests/Dumper/PhpDumperTest.php
@@ -1976,6 +1976,110 @@ PHP
         $this->assertInstanceOf(SingleMethodInterface::class, $container->get('bar')->foo);
         $this->assertInstanceOf(Foo::class, $container->get('bar')->foo->theMethod());
     }
+
+    /**
+     * @dataProvider getStripCommentsCodes
+     */
+    public function testStripComments(string $source, string $expected)
+    {
+        $reflection = new \ReflectionClass(PhpDumper::class);
+        $method = $reflection->getMethod('stripComments');
+
+        $output = $method->invoke(null, $source);
+
+        // Heredocs are preserved, making the output mixing Unix and Windows line
+        // endings, switching to "\n" everywhere on Windows to avoid failure.
+        if ('\\' === \DIRECTORY_SEPARATOR) {
+            $expected = str_replace("\r\n", "\n", $expected);
+            $output = str_replace("\r\n", "\n", $output);
+        }
+
+        $this->assertEquals($expected, $output);
+    }
+
+    public static function getStripCommentsCodes(): array
+    {
+        return [
+            ['<?php echo foo();', '<?php echo foo();'],
+            ['<?php echo/**/foo();', '<?php echo foo();'],
+            ['<?php echo/** bar */foo();', '<?php echo foo();'],
+            ['<?php /**/echo foo();', '<?php echo foo();'],
+            ['<?php echo \foo();', '<?php echo \foo();'],
+            ['<?php echo/**/\foo();', '<?php echo \foo();'],
+            ['<?php echo/** bar */\foo();', '<?php echo \foo();'],
+            ['<?php /**/echo \foo();', '<?php echo \foo();'],
+            [<<<'EOF'
+<?php
+include_once \dirname(__DIR__).'/foo.php';
+
+$string = 'string should not be   modified';
+
+$string = 'string should not be
+
+modified';
+
+
+$heredoc = <<<HD
+
+
+Heredoc should not be   modified {$a[1+$b]}
+
+
+HD;
+
+$nowdoc = <<<'ND'
+
+
+Nowdoc should not be   modified
+
+
+ND;
+
+/**
+ * some class comments to strip
+ */
+class TestClass
+{
+    /**
+     * some method comments to strip
+     */
+    public function doStuff()
+    {
+        // inline comment
+    }
+}
+EOF
+                , <<<'EOF'
+<?php
+include_once \dirname(__DIR__).'/foo.php';
+$string = 'string should not be   modified';
+$string = 'string should not be
+
+modified';
+$heredoc = <<<HD
+
+
+Heredoc should not be   modified {$a[1+$b]}
+
+
+HD;
+$nowdoc = <<<'ND'
+
+
+Nowdoc should not be   modified
+
+
+ND;
+class TestClass
+{
+    public function doStuff()
+    {
+        }
+}
+EOF
+            ],
+        ];
+    }
 }
 
 class Rot13EnvVarProcessor implements EnvVarProcessorInterface